New comments fixed
